pyproject.toml
[project]
name = "PI-BackEnd-DjangoDRF"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Caussz", email = "cczcaua@gmail.com"},
]
dependencies = ["django>=5.1.4", "django-rest-framework>=0.1.0", "djangorestframework>=3.15.2", "django-cors-headers>=4.6.0", "djangorestframework-simplejwt>=5.3.1", "django-rest-swagger>=2.2.0", "drf-spectacular>=0.28.0", "requests>=2.32.3"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python manage.py runserver 0.0.0.0:19003"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
post_migrate = "python manage.py graph_models --disable-sort-fields -g -o core.png core"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
test = "python manage.py test"

[tool.pdm]
distribution = false
